Split client code base into different components
