+++ app/javascript/src/components/Clients/constants.ts
@@ -1,0 +1,48 @@
+export const tableHeader = [
+  {
+    Header: "CLIENT",
+    accessor: "col1", // accessor is the "key" in the data
+    cssClass: "md:w-1/3 capitalize",
+  },
+  {
+    Header: "EMAIL ID",
+    accessor: "col2",
+    cssClass: "md:w-1/3",
+  },
+  {
+    Header: "HOURS LOGGED",
+    accessor: "col3",
+    cssClass: "text-right md:w-1/5", // accessor is the "key" in the data
+  },
+];
+
+export const employeeTableHeader = [
+  {
+    Header: "CLIENT",
+    accessor: "col1", // accessor is the "key" in the data
+    cssClass: "",
+  },
+  {
+    Header: "HOURS LOGGED",
+    accessor: "col3",
+    cssClass: "text-right", // accessor is the "key" in the data
+  },
+];
+
+export const mobileTableHeader = [
+  {
+    Header: "CLIENT",
+    accessor: "col1", // accessor is the "key" in the data
+    cssClass: "table__header font-medium",
+  },
+  {
+    Header: "HOURS",
+    accessor: "col3",
+    cssClass: "table__header font-medium text-right", // accessor is the "key" in the data
+  },
+  {
+    Header: "",
+    accessor: "col4",
+    cssClass: "font-medium text-right", // accessor is the "key" in the data
+  },
+];
